Flight schedule update complete

+++ Aeroclub.Cargo/Aeroclub.Cargo.Application/Services/FlightScheduleManagementService.cs
@@ -15,6 +15,7 @@ using AutoMapper;
 using Aeroclub.Cargo.Infrastructure.DateGenerator.Interfaces;
 using Aeroclub.Cargo.Infrastructure.DateGenerator.Models;
 using static Grpc.Core.Metadata;
+using Aeroclub.Cargo.Application.Models.Queries.FlightScheduleQMs;
 
 namespace Aeroclub.Cargo.Application.Services
 {
@@ -24,12 +25,14 @@ namespace Aeroclub.Cargo.Application.Services
         private readonly IDateGeneratorService _dateGeneratorService;
         private readonly IFlightService _flightService;
         private readonly ISectorService _sectorService;
+        private readonly ILayoutCloneService _layoutCloneService;
 
         public FlightScheduleManagementService(
             IUnitOfWork unitOfWork,
             IFlightService flightService,
             ISectorService sectorService,
             IDateGeneratorService dateGeneratorService,
+            ILayoutCloneService layoutCloneService,
             IMapper mapper, IFlightScheduleService flightScheduleService) :
             base(unitOfWork, mapper)
         {
@@ -37,6 +40,7 @@ namespace Aeroclub.Cargo.Application.Services
             _flightService = flightService;
             _sectorService = sectorService;
             _dateGeneratorService = dateGeneratorService;
+            _layoutCloneService= layoutCloneService;
         }
 
 
@@ -69,33 +73,40 @@ namespace Aeroclub.Cargo.Application.Services
 
         public async Task<ServiceResponseCreateStatus> UpdateAsync(FlightScheduleManagementUpdateRM dto)
         {
-            var res = new ServiceResponseCreateStatus();
+            var existingScheduleDeleteResponse = await DeleteFlightSchedule(dto.FlightScheduleIds);
+
+            if (existingScheduleDeleteResponse.StatusCode == ServiceResponseStatus.Failed || existingScheduleDeleteResponse.StatusCode == ServiceResponseStatus.ValidationError)
+            {
+                return existingScheduleDeleteResponse;
+            }
+
             var flightScheduleManagementEntity = _mapper.Map<FlightScheduleManagement>(dto);
 
             _unitOfWork.Repository<FlightScheduleManagement>().Update(flightScheduleManagementEntity);
             await _unitOfWork.SaveChangesAsync();
             _unitOfWork.Repository<FlightScheduleManagement>().Detach(flightScheduleManagementEntity);
 
+            var flightScheduleResponse = await CreateFlightSchedule(dto, flightScheduleManagementEntity.Id);
+            if (flightScheduleResponse.StatusCode == ServiceResponseStatus.Failed || flightScheduleResponse.StatusCode == ServiceResponseStatus.ValidationError)
+            {
+                await DeleteAsync(flightScheduleManagementResponse.Id);
+            }
 
-          /*if (flightScheduleManagementResponse == null)
+            if (flightScheduleManagementResponse == null)
             {
                 res.StatusCode = ServiceResponseStatus.Failed;
             }
             else
             {
-                var flightScheduleResponse = await CreateFlightSchedule(dto, flightScheduleManagementResponse.Id);
-                if (flightScheduleResponse.StatusCode == ServiceResponseStatus.Failed || flightScheduleResponse.StatusCode == ServiceResponseStatus.ValidationError)
-                {
-                    await DeleteAsync(flightScheduleManagementResponse.Id);
-                }
+
 
                 res = flightScheduleResponse;
-            }*/
+            }
 
             return res;
         }
 
-
+       
         public async Task<FlightScheduleManagementVM> GetAsync(FlightScheduleManagementDetailQM query)
         {
             var spec = new FlightScheduleManagementSpecification(query);
@@ -205,6 +216,37 @@ namespace Aeroclub.Cargo.Application.Services
             else
                 return new ServiceResponseCreateStatus() { StatusCode = ServiceResponseStatus.ValidationError };
 
+            return new ServiceResponseCreateStatus() { StatusCode = ServiceResponseStatus.Success };
+        }
+
+        private async Task<ServiceResponseCreateStatus> DeleteFlightSchedule(List<Guid> flightScheduleIds)
+        {
+            if(flightScheduleIds.Count() <1) return new ServiceResponseCreateStatus() { StatusCode = ServiceResponseStatus.ValidationError };
+            using (var transaction = _unitOfWork.BeginTransaction())
+            {
+                foreach (var scheduleId in flightScheduleIds)
+                {
+                    var spec = new FlightScheduleSpecification(new FlightScheduleQM() { Id = scheduleId, IncludeFlightScheduleSectors = true });
+                    var flightSchedule = await _unitOfWork.Repository<FlightSchedule>().GetEntityWithSpecAsync(spec);
+                    if (flightSchedule == null || flightSchedule.FlightScheduleSectors == null || flightSchedule.FlightScheduleSectors.Count() < 1) {
+                        transaction.Rollback();
+                        return new ServiceResponseCreateStatus() { StatusCode = ServiceResponseStatus.ValidationError };
+                    }
+
+                    var deleteClonedLayoutResponse = await _layoutCloneService.DeleteClonedCargoLayoutAsync(flightSchedule);
+                    if (!deleteClonedLayoutResponse) {
+                        transaction.Rollback();
+                        return new ServiceResponseCreateStatus() { StatusCode = ServiceResponseStatus.Failed }; 
+                    }
+
+                    var flightScheduleDeleteResponse = await _flightScheduleService.DeleteAsync(scheduleId);
+                    if (!flightScheduleDeleteResponse) {
+                        transaction.Rollback();
+                        return new ServiceResponseCreateStatus() { StatusCode = ServiceResponseStatus.Failed }; 
+                    }
+                }
+                transaction.Commit();
+            }
             return new ServiceResponseCreateStatus() { StatusCode = ServiceResponseStatus.Success };
         }
